Улучшена работа с городами и местами
Resolve https://github.com/stas12312/my-tickets-bot/issues/97
