+++ my_tickets_bot/src/bot/buttons.py
@@ -11,9 +11,10 @@ class MainMenu(StrEnum):
 
 class Settings(StrEnum):
     """Настройки"""
-    MY_CITIES = '🏙 Мои города'
+    CITIES_AND_PLACES = '🏙 Города и места'
     ADD_CITY = '➕ Добавить город'
     DELETE_CITY = '🗑 Удалить город'
+    PLACES = '🎭 Места'
 
     MY_PLACES = '🎭 Мои места'
     ADD_LOCATION = '➕ Добавить место'
