+++ my_tickets_bot/src/bot/handlers/__init__.py
@@ -1,10 +1,7 @@
-"""Обработчики"""
 from aiogram import Router
 
-from . import events
-from .cities import cities_handler
+from . import cities, locations, events
 from .common import common_handlers
-from .location import locations_handler
 from .settings import settings_router
 from .ticket import tickets_handler
 
@@ -14,6 +11,6 @@ main_router = Router()
 main_router.include_router(common_handlers)
 main_router.include_router(settings_router)
 main_router.include_router(events.router)
-main_router.include_router(cities_handler)
-main_router.include_router(locations_handler)
+main_router.include_router(cities.router)
+main_router.include_router(locations.router)
 main_router.include_router(tickets_handler)
