+++ my_tickets_bot/src/bot/handlers/cities/__init__.py
@@ -1,0 +1,9 @@
+"""Обработчики для городов"""
+from aiogram import Router
+
+from . import add, edit, show
+
+router = Router()
+router.include_router(add.router)
+router.include_router(edit.router)
+router.include_router(show.router)
