+++ my_tickets_bot/src/models/location.py
@@ -11,7 +11,3 @@ class Location:
     name: str | None = None
     address: str | None = None
     city: City | None = None
-
-    def get_show_text(self) -> str:
-        """Формирования строки для отображения"""
-        return f'{self.name} ({self.city.name}, {self.address})'
