+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["Stanislav Rush <911rush@gmail.com>"]
 
 [tool.poetry.dependencies]
 python = "^3.11"
-aiogram = "^3.0.0b6"
+aiogram = "^3.0.0b7"
 asyncpg = "^0.27.0"
 alembic = "^1.9.3"
 tzfpy = "^0.13.1"
